Updates updates and additions (issue too)
Added new utilities to complete an action when something is a value or returns a value which is threaded so that it can be continually checked in the background. Engine also is now able to take screenshots and entities are able to have the clickability set. Textfield now takes reference from the global package constants. Issue added on github (to do with image objects and opacity)